Bump source/binary level to 11, drop MR-jar
